Fixed the inability to delete custom flairs with relative URLs or special characters within their URL. http://phistuck-app.appspot.com/view-bug?extension=Flair%20Extender%20For%20Google%20Calendar&key=agxwaGlzdHVjay1hcHByFQsSDUlzc3VlRGF0YWJhc2UYqYwVDA
The "src" property of the image element returned a full URL instead of the actual string that was set, so the custom flair was never found within the actual list and could not be removed. A data property was added and used for the comparison, which keeps the string intact.